+++ babel.config.js
@@ -2,6 +2,7 @@ module.exports = function(api) {
 	api.cache(true);
 	return {
 		presets: ['babel-preset-expo', '@bluebase/code-standards/babel.config'],
+		plugins: ['react-native-reanimated/plugin'],
 		env: {
 			test: {
 				plugins: [