+++ jest.config.js
@@ -18,4 +18,12 @@ const modules = [
 module.exports = Object.assign(configs, {
 	transformIgnorePatterns: [`/node_modules/(?!${modules})`],
 	setupFilesAfterEnv: ['<rootDir>/tests/setup.ts'],
+	coverageThreshold: {
+		global: {
+			branches: 75,
+			functions: 75,
+			lines: 75,
+			statements: 75
+		}
+	},
 });