+++ plugins.ts
@@ -1,7 +1,7 @@
 import BlueBasePluginJsonSchemaComponents from '@bluebase/plugin-json-schema-components';
 import BlueBasePluginLauncher from '@bluebase/plugin-launcher';
 import BlueBasePluginMaterialUI from '@bluebase/plugin-material-ui';
-import BlueBasePluginReactRouter from '@bluebase/plugin-react-router';
+import BlueBasePluginReactNavigation from '@bluebase/plugin-react-navigation';
 import BlueBasePluginResponsiveGrid from '@bluebase/plugin-responsive-grid';
 import { MaterialCommunityIcons } from '@bluebase/plugin-vector-icons';
 
@@ -11,7 +11,7 @@ export const plugins = [
 	BlueBasePluginJsonSchemaComponents,
 	BlueBasePluginLauncher,
 	BlueBasePluginMaterialUI,
-	BlueBasePluginReactRouter,
+	BlueBasePluginReactNavigation,
 	BlueBasePluginResponsiveGrid,
 	MaterialCommunityIcons,
 	Plugin,