+++ src/components/TermsOfService/TermsOfService.tsx
@@ -1,8 +1,16 @@
 import { Text } from '@bluebase/components';
-import { useIntl } from '@bluebase/core';
-import React from 'react';
+import { useIntl, useNavigation } from '@bluebase/core';
+import React, { useLayoutEffect } from 'react';
 
 export const TermsOfService = () => {
 	const { __ } = useIntl();
+	const { setOptions } = useNavigation();
+
+	const title = __('Terms of Service');
+
+	useLayoutEffect(() => {
+		setOptions({ title });
+	}, [title]);
+
 	return <Text>{__('Please add terms and conditions here')}</Text>;
 };