+++ src/index.ts
@@ -4,12 +4,16 @@ import { components } from './components';
 import { defaultConfigs } from './configs';
 import { lang } from './lang';
 import { layouts } from './layouts';
+import { SettingsPageItem } from './layouts/SettingsPageItem';
 import { createSettingsRoutes } from './lib';
+import { navigation } from './navigation';
 import { routes } from './routes';
 import { settings } from './settings';
 import { VERSION } from './version';
+
 export * from './components/exports';
 export { SettingsPageProps } from './layouts/SettingsPage';
+export * from './layouts/SettingsPageItem';
 export { createSettingsRoutes };
 
 export default createPlugin({
@@ -29,6 +33,8 @@ export default createPlugin({
 		...components,
 		...settings,
 		...layouts,
+
+		SettingsPageItem: SettingsPageItem,
 	},
 
 	indexRoute: 'SettingsApp',
@@ -37,5 +43,6 @@ export default createPlugin({
 
 	filters: {
 		...lang,
+		...navigation,
 	},
 });