+++ src/layouts/SettingsLayoutDesktop/SettingsLayoutDesktop.tsx
@@ -1,9 +1,18 @@
-import { Column, Container, Row } from '@bluebase/components';
+import { SafeAreaView, ScrollView, View } from '@bluebase/components';
+import { useIntl, useStyles, useTheme } from '@bluebase/core';
 import React from 'react';
+import { ViewStyle } from 'react-native';
 
 import { CreateSettingsRoutesOptions } from '../../lib';
 import { SettingsPageDesktop } from '../SettingsPage';
 import { SettingsPageList } from '../SettingsPageList';
+
+export interface SettingsLayoutDesktopStyles {
+	root: ViewStyle;
+	menuColumn: ViewStyle;
+	contentColumn: ViewStyle;
+	contentWrapper: ViewStyle;
+}
 
 export interface SettingsLayoutDesktopProps extends CreateSettingsRoutesOptions {
 	page?: string;
@@ -11,20 +20,48 @@ export interface SettingsLayoutDesktopProps extends CreateSettingsRoutesOptions 
 
 export const SettingsLayoutDesktop = (props: SettingsLayoutDesktopProps) => {
 	const { pages, page, filter, mainRoute } = props;
+	const { theme } = useTheme();
+	const { rtl } = useIntl();
 
 	const pageObj = pages.find(p => p.name === page);
 
+	const styles = useStyles<SettingsLayoutDesktopStyles>('SettingsLayoutDesktop', props, {
+		root: {
+			flex: 1,
+			flexDirection: 'row',
+		},
+		menuColumn: {
+			backgroundColor: theme.palette.background.light,
+			width: 275,
+			borderRightWidth: rtl ? 0 : 1,
+			borderLeftWidth: rtl ? 1 : 0,
+			borderColor: theme.palette.divider,
+		},
+		contentColumn: {
+			flexGrow: 1,
+		},
+		contentWrapper: {
+			alignSelf: 'center',
+			marginHorizontal: theme.spacing.unit * 2,
+			width: 600,
+		}
+	});
+
 	return (
-		<Container>
-			<Row>
-				<Column xl={2.5} lg={3} md={3} sm={12} xs={12}>
-					<SettingsPageList pages={pages} name={mainRoute.name} />
-				</Column>
-				<Column xl={9.5} lg={9} md={9} sm={12} xs={12}>
+		<View style={styles.root}>
+			<View style={styles.menuColumn}>
+				<ScrollView>
+					<SafeAreaView>
+						<SettingsPageList pages={pages} name={mainRoute.name} />
+					</SafeAreaView>
+				</ScrollView>
+			</View>
+			<ScrollView style={styles.contentColumn} contentContainerStyle={styles.contentWrapper}>
+				<SafeAreaView>
 					{pageObj ? <SettingsPageDesktop filter={filter} {...pageObj} {...props} /> : null}
-				</Column>
-			</Row>
-		</Container>
+				</SafeAreaView>
+			</ScrollView>
+		</View>
 	);
 };