+++ src/layouts/SettingsPage/SettingsPageDesktop.tsx
@@ -1,44 +1,51 @@
-import { H6, NavigationOptions, View } from '@bluebase/components';
-import {
-	resolveThunk,
-	Theme,
-	useFilter,
-	useIntl,
-	useStyles
-} from '@bluebase/core';
-import get from 'lodash.get';
+import { H6, StackNavigationOptions, View } from '@bluebase/components';
+import { Theme, useFilter, useIntl, useStyles } from '@bluebase/core';
 import React from 'react';
-import { Dimensions, ScrollView, TextStyle, ViewStyle } from 'react-native';
+import { TextStyle, ViewStyle } from 'react-native';
 
-import { useScreenProps } from '../../hooks';
 import { SettingsPageProps } from '../SettingsPage';
-import { SettingsPageItemDesktop } from '../SettingsPageItem';
+import { SettingsPageItemDesktop, SettingsPageItemDesktopStyles } from '../SettingsPageItem';
 
 export interface SettingsPageDesktopStyles {
+	header: ViewStyle;
+	headerRight: ViewStyle;
+	headerTitle: ViewStyle;
 	title: TextStyle;
 	root: ViewStyle;
-	scrollViewContainer: ViewStyle;
+	item: Partial<SettingsPageItemDesktopStyles>;
 }
 
-export interface SettingsPageDesktopProps extends SettingsPageProps {
+export type SettingsPageDesktopProps = SettingsPageProps & {
 	styles?: Partial<SettingsPageDesktopStyles>;
-}
+};
 
 const defaultStyles = (theme: Theme): SettingsPageDesktopStyles => {
-	const { height } = Dimensions.get('window');
+	// const { height } = Dimensions.get('window');
 
 	return {
 		root: {
 			flex: 1,
 		},
-		scrollViewContainer: {
-			height: height - theme.spacing.unit * 10,
+		header: {
+			flexDirection: 'row',
+			marginVertical: theme.spacing.unit,
+		},
+		headerTitle: {
+			flexGrow: 1
+		},
+		headerRight: {
+			justifyContent: 'center',
 		},
 		title: {
 			paddingHorizontal: theme.spacing.unit * 2,
 			paddingTop: theme.spacing.unit * 2,
 			paddingBottom: theme.spacing.unit,
 		},
+		item: {
+			root: {
+				marginBottom: theme.spacing.unit * 2,
+			}
+		}
 	};
 };
 
@@ -46,30 +53,29 @@ export const SettingsPageDesktop = (props: SettingsPageDesktopProps) => {
 	const { filter, items } = props;
 	const { __ } = useIntl();
 
-	const screenProps = useScreenProps();
 	const styles = useStyles('SettingsPageDesktop', props, defaultStyles);
 	const { value: filteredItems } = useFilter(`${filter}.page.desktop`, items, props);
 
-	const options: NavigationOptions = resolveThunk(get(props, 'options', {}), screenProps);
-	const title = get(options, 'title', options.headerTitle) as string;
+	const { title, headerRight }: StackNavigationOptions = props.options || {};
 
 	return (
-		<ScrollView
-			style={{ flex: 1 }}
-			contentContainerStyle={styles.scrollViewContainer}
-			showsVerticalScrollIndicator={false}
-		>
-			<View style={styles.root}>
-				{title && <H6 style={styles.title}>{__(title)}</H6>}
-				{filteredItems.map(item => (
-					<SettingsPageItemDesktop
-						key={item.name}
-						{...item}
-						styles={{ root: { marginVertical: 8 } }}
-					/>
-				))}
+		<View style={styles.root}>
+			<View style={styles.header}>
+				<View style={styles.headerTitle}>
+					{title && <H6 style={styles.title}>{__(title)}</H6>}
+				</View>
+				<View style={styles.headerRight}>
+					{headerRight ? headerRight({}) : null}
+				</View>
 			</View>
-		</ScrollView>
+			{filteredItems.map(item => (
+				<SettingsPageItemDesktop
+					key={item.name}
+					{...item}
+					styles={styles.item}
+				/>
+			))}
+		</View>
 	);
 };