+++ src/settings/TermsOfServiceSetting/TermsOfServiceSetting.tsx
@@ -1,33 +1,20 @@
-import { Dialog, Divider, List, View } from '@bluebase/components';
-import { getComponent, useIntl, useTheme } from '@bluebase/core';
-import React, { useState } from 'react';
-
-const TermsOfService = getComponent('TermsOfService', 'EmptyState');
+import { List } from '@bluebase/components';
+import { useIntl, useNavigation } from '@bluebase/core';
+import React, { useCallback } from 'react';
 
 export const TermsOfServiceSetting = () => {
 	const { __ } = useIntl();
-	const { theme } = useTheme();
+	const { navigate } = useNavigation();
 
-	const [visible, setVisible] = useState(false);
-
-	const toggle = () => setVisible(!visible);
+	const goToTermsOfService = useCallback(() => {
+		navigate('TermsOfService');
+	}, [navigate]);
 
 	return (
-		<React.Fragment>
-			<List.Item
-				left={<List.Icon name="file-document" />}
-				title={__('Terms of Service')}
-				onPress={toggle}
-			/>
-			<Dialog dismissable visible={visible} onDismiss={toggle}>
-				<View testID="test-listItem">
-					<List.Subheader>{__('Terms of Service')}</List.Subheader>
-					<Divider />
-					<View style={{ padding: theme.spacing.unit * 2 }}>
-						<TermsOfService />
-					</View>
-				</View>
-			</Dialog>
-		</React.Fragment>
+		<List.Item
+			left={<List.Icon name="file-document" />}
+			title={__('Terms of Service')}
+			onPress={goToTermsOfService}
+		/>
 	);
 };